apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
pgvector==0.3.2
alembic==1.13.2
pydantic==2.8.2
pydantic-settings==2.4.0
python-multipart==0.0.9

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-auth==2.35.0

# AI
anthropic==0.36.0
openai==1.51.0

# Document parsing
pdfplumber==0.11.4
python-docx==1.1.2

# Task queue
celery==5.4.0
redis==5.0.8
flower==2.0.1

# Billing
stripe==10.12.0
razorpay==1.4.1

# Email
resend==2.3.0

# HTTP client
httpx==0.27.2
